@@ -19,7 +19,6 @@
 
 namespace Doctrine\Instantiator;
 
-use Closure;
 use Doctrine\Instantiator\Exception\InvalidArgumentException;
 use Doctrine\Instantiator\Exception\UnexpectedValueException;
 use Exception;
@@ -41,14 +40,14 @@ final class Instantiator implements InstantiatorInterface
     const SERIALIZATION_FORMAT_AVOID_UNSERIALIZER = 'O';
 
     /**
-     * @var \Closure[] of {@see \Closure} instances used to instantiate specific classes
+     * @var \callable[] used to instantiate specific classes, indexed by class name
      */
-    private static $cachedInstantiators = array();
+    private static $cachedInstantiators = [];
 
     /**
-     * @var object[] of objects that can directly be cloned
+     * @var object[] of objects that can directly be cloned, indexed by class name
      */
-    private static $cachedCloneables = array();
+    private static $cachedCloneables = [];
 
     /**
      * {@inheritDoc}
@@ -71,11 +70,9 @@ final class Instantiator implements InstantiatorInterface
     /**
      * Builds the requested object and caches it in static properties for performance
      *
-     * @param string $className
-     *
      * @return object
      */
-    private function buildAndCacheFromFactory($className)
+    private function buildAndCacheFromFactory(string $className)
     {
         $factory  = self::$cachedInstantiators[$className] = $this->buildFactory($className);
         $instance = $factory();
@@ -88,26 +85,24 @@ final class Instantiator implements InstantiatorInterface
     }
 
     /**
-     * Builds a {@see \Closure} capable of instantiating the given $className without
+     * Builds a callable capable of instantiating the given $className without
      * invoking its constructor.
      *
-     * @param string $className
-     *
-     * @return Closure
+     * @throws InvalidArgumentException
+     * @throws UnexpectedValueException
+     * @throws \ReflectionException
      */
-    private function buildFactory($className)
+    private function buildFactory(string $className) : callable
     {
         $reflectionClass = $this->getReflectionClass($className);
 
         if ($this->isInstantiableViaReflection($reflectionClass)) {
-            return function () use ($reflectionClass) {
-                return $reflectionClass->newInstanceWithoutConstructor();
-            };
+            return [$reflectionClass, 'newInstanceWithoutConstructor'];
         }
 
         $serializedString = sprintf(
             '%s:%d:"%s":0:{}',
-            $this->getSerializationFormat($reflectionClass),
+            self::SERIALIZATION_FORMAT_AVOID_UNSERIALIZER,
             strlen($className),
             $className
         );
@@ -125,8 +120,9 @@ final class Instantiator implements InstantiatorInterface
      * @return ReflectionClass
      *
      * @throws InvalidArgumentException
+     * @throws \ReflectionException
      */
-    private function getReflectionClass($className)
+    private function getReflectionClass($className) : ReflectionClass
     {
         if (! class_exists($className)) {
             throw InvalidArgumentException::fromNonExistingClass($className);
@@ -149,9 +145,9 @@ final class Instantiator implements InstantiatorInterface
      *
      * @return void
      */
-    private function checkIfUnSerializationIsSupported(ReflectionClass $reflectionClass, $serializedString)
+    private function checkIfUnSerializationIsSupported(ReflectionClass $reflectionClass, $serializedString) : void
     {
-        set_error_handler(function ($code, $message, $file, $line) use ($reflectionClass, & $error) {
+        set_error_handler(function ($code, $message, $file, $line) use ($reflectionClass, & $error) : void {
             $error = UnexpectedValueException::fromUncleanUnSerialization(
                 $reflectionClass,
                 $message,
@@ -178,7 +174,7 @@ final class Instantiator implements InstantiatorInterface
      *
      * @return void
      */
-    private function attemptInstantiationViaUnSerialization(ReflectionClass $reflectionClass, $serializedString)
+    private function attemptInstantiationViaUnSerialization(ReflectionClass $reflectionClass, $serializedString) : void
     {
         try {
             unserialize($serializedString);
@@ -189,28 +185,15 @@ final class Instantiator implements InstantiatorInterface
         }
     }
 
-    /**
-     * @param ReflectionClass $reflectionClass
-     *
-     * @return bool
-     */
-    private function isInstantiableViaReflection(ReflectionClass $reflectionClass)
+    private function isInstantiableViaReflection(ReflectionClass $reflectionClass) : bool
     {
-        if (\PHP_VERSION_ID >= 50600) {
-            return ! ($this->hasInternalAncestors($reflectionClass) && $reflectionClass->isFinal());
-        }
-
-        return \PHP_VERSION_ID >= 50400 && ! $this->hasInternalAncestors($reflectionClass);
+        return ! ($this->hasInternalAncestors($reflectionClass) && $reflectionClass->isFinal());
     }
 
     /**
      * Verifies whether the given class is to be considered internal
-     *
-     * @param ReflectionClass $reflectionClass
-     *
-     * @return bool
      */
-    private function hasInternalAncestors(ReflectionClass $reflectionClass)
+    private function hasInternalAncestors(ReflectionClass $reflectionClass) : bool
     {
         do {
             if ($reflectionClass->isInternal()) {
@@ -222,52 +205,12 @@ final class Instantiator implements InstantiatorInterface
     }
 
     /**
-     * Verifies if the given PHP version implements the `Serializable` interface serialization
-     * with an incompatible serialization format. If that's the case, use serialization marker
-     * "C" instead of "O".
-     *
-     * @link http://news.php.net/php.internals/74654
-     *
-     * @param ReflectionClass $reflectionClass
-     *
-     * @return string the serialization format marker, either self::SERIALIZATION_FORMAT_USE_UNSERIALIZER
-     *                or self::SERIALIZATION_FORMAT_AVOID_UNSERIALIZER
-     */
-    private function getSerializationFormat(ReflectionClass $reflectionClass)
-    {
-        if ($this->isPhpVersionWithBrokenSerializationFormat()
-            && $reflectionClass->implementsInterface('Serializable')
-        ) {
-            return self::SERIALIZATION_FORMAT_USE_UNSERIALIZER;
-        }
-
-        return self::SERIALIZATION_FORMAT_AVOID_UNSERIALIZER;
-    }
-
-    /**
-     * Checks whether the current PHP runtime uses an incompatible serialization format
-     *
-     * @return bool
-     */
-    private function isPhpVersionWithBrokenSerializationFormat()
-    {
-        return PHP_VERSION_ID === 50429 || PHP_VERSION_ID === 50513;
-    }
-
-    /**
      * Checks if a class is cloneable
      *
-     * @param ReflectionClass $reflection
-     *
-     * @return bool
+     * Classes implementing `__clone` cannot be safely cloned, as that may cause side-effects.
      */
-    private function isSafeToClone(ReflectionClass $reflection)
+    private function isSafeToClone(ReflectionClass $reflection) : bool
     {
-        if (method_exists($reflection, 'isCloneable') && ! $reflection->isCloneable()) {
-            return false;
-        }
-
-        // not cloneable if it implements `__clone`, as we want to avoid calling it
-        return ! $reflection->hasMethod('__clone');
+        return $reflection->isCloneable() && ! $reflection->hasMethod('__clone');
     }
 }
