@@ -166,7 +166,7 @@ final class StandardTagFactory implements TagFactory
      */
     private function extractTagParts($tagLine)
     {
-        $matches = array();
+        $matches = [];
         if (! preg_match('/^@(' . self::REGEX_TAGNAME . ')(?:\s*([^\s].*)|$)/us', $tagLine, $matches)) {
             throw new \InvalidArgumentException(
                 'The tag "' . $tagLine . '" does not seem to be wellformed, please check it for errors'
@@ -196,8 +196,7 @@ final class StandardTagFactory implements TagFactory
         $arguments        = $this->getArgumentsForParametersFromWiring(
             $this->fetchParametersForHandlerFactoryMethod($handlerClassName),
             $this->getServiceLocatorWithDynamicParameters($context, $name, $body)
-        )
-        ;
+        );
 
         return call_user_func_array([$handlerClassName, 'create'], $arguments);
     }
