@@ -93,7 +93,7 @@ final class DocBlockFactory implements DocBlockFactoryInterface
         return new DocBlock(
             $summary,
             $description ? $this->descriptionFactory->create($description, $context) : null,
-            array_filter($this->parseTagBlock($tags, $context), function($tag) {
+            array_filter($this->parseTagBlock($tags, $context), function ($tag) {
                 return $tag instanceof Tag;
             }),
             $context,
@@ -120,11 +120,11 @@ final class DocBlockFactory implements DocBlockFactoryInterface
         $comment = trim(preg_replace('#[ \t]*(?:\/\*\*|\*\/|\*)?[ \t]{0,1}(.*)?#u', '$1', $comment));
 
         // reg ex above is not able to remove */ from a single line docblock
-        if (substr($comment, -2) == '*/') {
+        if (substr($comment, -2) === '*/') {
             $comment = trim(substr($comment, 0, -2));
         }
 
-        return str_replace(array("\r\n", "\r"), "\n", $comment);
+        return str_replace(["\r\n", "\r"], "\n", $comment);
     }
 
     /**
@@ -143,7 +143,7 @@ final class DocBlockFactory implements DocBlockFactoryInterface
         // method does not split tags so we return this verbatim as the fourth result (tags). This saves us the
         // performance impact of running a regular expression
         if (strpos($comment, '@') === 0) {
-            return array('', '', '', $comment);
+            return ['', '', '', $comment];
         }
 
         // clears all extra horizontal whitespace from the line endings to prevent parsing issues
@@ -241,7 +241,7 @@ final class DocBlockFactory implements DocBlockFactoryInterface
      */
     private function splitTagBlockIntoTagLines($tags)
     {
-        $result = array();
+        $result = [];
         foreach (explode("\n", $tags) as $tag_line) {
             if (isset($tag_line[0]) && ($tag_line[0] === '@')) {
                 $result[] = $tag_line;
