@@ -1,6 +1,6 @@
 <?php
 /*
- * This file is part of the PHP_TokenStream package.
+ * This file is part of php-token-stream.
  *
  * (c) Sebastian Bergmann <sebastian@phpunit.de>
  *
@@ -10,22 +10,17 @@
 
 /**
  * A caching factory for token stream objects.
- *
- * @author    Sebastian Bergmann <sebastian@phpunit.de>
- * @copyright Sebastian Bergmann <sebastian@phpunit.de>
- * @license   http://www.opensource.org/licenses/BSD-3-Clause  The BSD 3-Clause License
- * @link      http://github.com/sebastianbergmann/php-token-stream/tree
- * @since     Class available since Release 1.0.0
  */
 class PHP_Token_Stream_CachingFactory
 {
     /**
      * @var array
      */
-    protected static $cache = array();
+    protected static $cache = [];
 
     /**
-     * @param  string $filename
+     * @param string $filename
+     *
      * @return PHP_Token_Stream
      */
     public static function get($filename)
@@ -45,7 +40,7 @@ class PHP_Token_Stream_CachingFactory
         if (is_string($filename)) {
             unset(self::$cache[$filename]);
         } else {
-            self::$cache = array();
+            self::$cache = [];
         }
     }
 }
