@@ -1,6 +1,6 @@
 <?php
 /*
- * This file is part of the PHP_TokenStream package.
+ * This file is part of php-token-stream.
  *
  * (c) Sebastian Bergmann <sebastian@phpunit.de>
  *
@@ -8,21 +8,14 @@
  * file that was distributed with this source code.
  */
 
-/**
- * Tests for the PHP_Token_FUNCTION class.
- *
- * @package    PHP_TokenStream
- * @subpackage Tests
- * @author     Sebastian Bergmann <sebastian@phpunit.de>
- * @copyright  Sebastian Bergmann <sebastian@phpunit.de>
- * @license    http://www.opensource.org/licenses/BSD-3-Clause  The BSD 3-Clause License
- * @version    Release: @package_version@
- * @link       http://github.com/sebastianbergmann/php-token-stream/
- * @since      Class available since Release 1.0.0
- */
-class PHP_Token_FunctionTest extends PHPUnit_Framework_TestCase
+use PHPUnit\Framework\TestCase;
+
+class PHP_Token_FunctionTest extends TestCase
 {
-    protected $functions;
+    /**
+     * @var PHP_Token_FUNCTION[]
+     */
+    private $functions;
 
     protected function setUp()
     {
@@ -40,23 +33,23 @@ class PHP_Token_FunctionTest extends PHPUnit_Framework_TestCase
      */
     public function testGetArguments()
     {
-        $this->assertEquals(array(), $this->functions[0]->getArguments());
+        $this->assertEquals([], $this->functions[0]->getArguments());
 
         $this->assertEquals(
-          array('$baz' => 'Baz'), $this->functions[1]->getArguments()
+          ['$baz' => 'Baz'], $this->functions[1]->getArguments()
         );
 
         $this->assertEquals(
-          array('$foobar' => 'Foobar'), $this->functions[2]->getArguments()
+          ['$foobar' => 'Foobar'], $this->functions[2]->getArguments()
         );
 
         $this->assertEquals(
-          array('$barfoo' => 'Barfoo'), $this->functions[3]->getArguments()
+          ['$barfoo' => 'Barfoo'], $this->functions[3]->getArguments()
         );
 
-        $this->assertEquals(array(), $this->functions[4]->getArguments());
+        $this->assertEquals([], $this->functions[4]->getArguments());
 
-        $this->assertEquals(array('$x' => null, '$y' => null), $this->functions[5]->getArguments());
+        $this->assertEquals(['$x' => null, '$y' => null], $this->functions[5]->getArguments());
     }
 
     /**
