@@ -1,6 +1,6 @@
 <?php
 /*
- * This file is part of the PHP_TokenStream package.
+ * This file is part of php-token-stream.
  *
  * (c) Sebastian Bergmann <sebastian@phpunit.de>
  *
@@ -8,22 +8,14 @@
  * file that was distributed with this source code.
  */
 
-/**
- * Tests for the PHP_Token_REQUIRE_ONCE, PHP_Token_REQUIRE
- * PHP_Token_INCLUDE_ONCE and PHP_Token_INCLUDE_ONCE classes.
- *
- * @package    PHP_TokenStream
- * @subpackage Tests
- * @author     Laurent Laville <pear@laurent-laville.org>
- * @copyright  Sebastian Bergmann <sebastian@phpunit.de>
- * @license    http://www.opensource.org/licenses/BSD-3-Clause  The BSD 3-Clause License
- * @version    Release: @package_version@
- * @link       http://github.com/sebastianbergmann/php-token-stream/
- * @since      Class available since Release 1.0.2
- */
-class PHP_Token_IncludeTest extends PHPUnit_Framework_TestCase
+use PHPUnit\Framework\TestCase;
+
+class PHP_Token_IncludeTest extends TestCase
 {
-    protected $ts;
+    /**
+     * @var PHP_Token_Stream
+     */
+    private $ts;
 
     protected function setUp()
     {
@@ -37,7 +29,7 @@ class PHP_Token_IncludeTest extends PHPUnit_Framework_TestCase
     public function testGetIncludes()
     {
         $this->assertSame(
-          array('test4.php', 'test3.php', 'test2.php', 'test1.php'),
+          ['test4.php', 'test3.php', 'test2.php', 'test1.php'],
           $this->ts->getIncludes()
         );
     }
@@ -49,13 +41,13 @@ class PHP_Token_IncludeTest extends PHPUnit_Framework_TestCase
     public function testGetIncludesCategorized()
     {
         $this->assertSame(
-          array(
-            'require_once' => array('test4.php'),
-            'require'      => array('test3.php'),
-            'include_once' => array('test2.php'),
-            'include'      => array('test1.php')
-          ),
-          $this->ts->getIncludes(TRUE)
+          [
+            'require_once' => ['test4.php'],
+            'require'      => ['test3.php'],
+            'include_once' => ['test2.php'],
+            'include'      => ['test1.php']
+          ],
+          $this->ts->getIncludes(true)
         );
     }
 
@@ -66,8 +58,8 @@ class PHP_Token_IncludeTest extends PHPUnit_Framework_TestCase
     public function testGetIncludesCategory()
     {
         $this->assertSame(
-          array('test4.php'),
-          $this->ts->getIncludes(TRUE, 'require_once')
+          ['test4.php'],
+          $this->ts->getIncludes(true, 'require_once')
         );
     }
 }
