@@ -70,7 +70,7 @@ class Var_ extends BaseTag implements Factory\StaticMethod
         }
 
         // if the next item starts with a $ or ...$ it must be the variable name
-        if (isset($parts[0]) && (strlen($parts[0]) > 0) && ($parts[0][0] == '$')) {
+        if (isset($parts[0]) && (strlen($parts[0]) > 0) && ($parts[0][0] === '$')) {
             $variableName = array_shift($parts);
             array_shift($parts);
 
@@ -111,8 +111,8 @@ class Var_ extends BaseTag implements Factory\StaticMethod
      */
     public function __toString()
     {
-        return ($this->type ? $this->type.' ' : '')
-            .(empty($this->variableName) ? null : ('$'.$this->variableName))
-            .($this->description ? ' '.$this->description : '');
+        return ($this->type ? $this->type . ' ' : '')
+            . (empty($this->variableName) ? null : ('$' . $this->variableName))
+            . ($this->description ? ' ' . $this->description : '');
     }
 }
