@@ -20,27 +20,17 @@ final class TaggedValue
     private $tag;
     private $value;
 
-    /**
-     * @param string $tag
-     * @param mixed  $value
-     */
-    public function __construct($tag, $value)
+    public function __construct(string $tag, $value)
     {
         $this->tag = $tag;
         $this->value = $value;
     }
 
-    /**
-     * @return string
-     */
-    public function getTag()
+    public function getTag(): string
     {
         return $this->tag;
     }
 
-    /**
-     * @return mixed
-     */
     public function getValue()
     {
         return $this->value;
