@@ -18,7 +18,7 @@ use Symfony\Component\Yaml\Exception\ParseException;
  *
  * @author Fabien Potencier <fabien@symfony.com>
  *
- * @final since version 3.4
+ * @final
  */
 class Yaml
 {
@@ -35,18 +35,12 @@ class Yaml
     const DUMP_EMPTY_ARRAY_AS_SEQUENCE = 1024;
 
     /**
-     * @deprecated since version 3.4, to be removed in 4.0. Quote your evaluable keys instead.
-     */
-    const PARSE_KEYS_AS_STRINGS = 2048;
-
-    /**
      * Parses a YAML file into a PHP value.
      *
-     *  Usage:
-     *  <code>
-     *   $array = Yaml::parseFile('config.yml');
-     *   print_r($array);
-     *  </code>
+     * Usage:
+     *
+     *     $array = Yaml::parseFile('config.yml');
+     *     print_r($array);
      *
      * @param string $filename The path to the YAML file to be parsed
      * @param int    $flags    A bit field of PARSE_* constants to customize the YAML parser behavior
@@ -55,7 +49,7 @@ class Yaml
      *
      * @throws ParseException If the file could not be read or the YAML is not valid
      */
-    public static function parseFile($filename, $flags = 0)
+    public static function parseFile(string $filename, int $flags = 0)
     {
         $yaml = new Parser();
 
@@ -78,34 +72,8 @@ class Yaml
      *
      * @throws ParseException If the YAML is not valid
      */
-    public static function parse($input, $flags = 0)
+    public static function parse(string $input, int $flags = 0)
     {
-        if (\is_bool($flags)) {
-            @trigger_error('Passing a boolean flag to toggle exception handling is deprecated since Symfony 3.1 and will be removed in 4.0. Use the PARSE_EXCEPTION_ON_INVALID_TYPE flag instead.', E_USER_DEPRECATED);
-
-            if ($flags) {
-                $flags = self::PARSE_EXCEPTION_ON_INVALID_TYPE;
-            } else {
-                $flags = 0;
-            }
-        }
-
-        if (\func_num_args() >= 3) {
-            @trigger_error('Passing a boolean flag to toggle object support is deprecated since Symfony 3.1 and will be removed in 4.0. Use the PARSE_OBJECT flag instead.', E_USER_DEPRECATED);
-
-            if (func_get_arg(2)) {
-                $flags |= self::PARSE_OBJECT;
-            }
-        }
-
-        if (\func_num_args() >= 4) {
-            @trigger_error('Passing a boolean flag to toggle object for map support is deprecated since Symfony 3.1 and will be removed in 4.0. Use the Yaml::PARSE_OBJECT_FOR_MAP flag instead.', E_USER_DEPRECATED);
-
-            if (func_get_arg(3)) {
-                $flags |= self::PARSE_OBJECT_FOR_MAP;
-            }
-        }
-
         $yaml = new Parser();
 
         return $yaml->parse($input, $flags);
@@ -124,26 +92,8 @@ class Yaml
      *
      * @return string A YAML string representing the original PHP value
      */
-    public static function dump($input, $inline = 2, $indent = 4, $flags = 0)
+    public static function dump($input, int $inline = 2, int $indent = 4, int $flags = 0): string
     {
-        if (\is_bool($flags)) {
-            @trigger_error('Passing a boolean flag to toggle exception handling is deprecated since Symfony 3.1 and will be removed in 4.0. Use the DUMP_EXCEPTION_ON_INVALID_TYPE flag instead.', E_USER_DEPRECATED);
-
-            if ($flags) {
-                $flags = self::DUMP_EXCEPTION_ON_INVALID_TYPE;
-            } else {
-                $flags = 0;
-            }
-        }
-
-        if (\func_num_args() >= 5) {
-            @trigger_error('Passing a boolean flag to toggle object support is deprecated since Symfony 3.1 and will be removed in 4.0. Use the DUMP_OBJECT flag instead.', E_USER_DEPRECATED);
-
-            if (func_get_arg(4)) {
-                $flags |= self::DUMP_OBJECT;
-            }
-        }
-
         $yaml = new Dumper($indent);
 
         return $yaml->dump($input, $inline, 0, $flags);
